moved composite_images to pype's tvpaint.lib
